+++ projects/hipcub/hipcub/include/hipcub/backend/rocprim/grid/grid_even_share.hpp
@@ -154,7 +154,7 @@ public:
         {
             // This thread block gets a normal share of grains (avg_tiles_per_block)
             block_offset = normal_base_offset + (block_id * normal_share_items);
-            block_end = _HIPCUB_STD::min(num_items, block_offset + normal_share_items);
+            block_end    = _HIPCUB_STD::min(num_items, block_offset + normal_share_items);
         }
         // Else default past-the-end
     }